requirements.txt
streamlit
pandas

pdfplumber
python-docx

langchain
langgraph
langsmith
langchain-openai
langchain-groq
python-dotenv
